requirements.txt
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
gunicorn==23.0.0
python-dotenv==1.0.1
